+++ cmd/aat/validate_cmd.go
@@ -227,6 +227,7 @@ func validateCommand(args *validateArgs, out io.Writer) int {
 			WithHeaderInputs(engine.TemplateHeaderInputs(g, registry)).
 			WithFormInputFields(engine.TemplateFormInputFields(g, registry)).
 			WithQueryInputParams(engine.TemplateQueryInputParams(g, registry)).
+			WithBodyInputFields(engine.TemplateBodyInputFields(g, registry)).
 			WithPathTemplates(engine.TemplatePaths(g, registry))
 	}
 	specPaths := validator.CollectSpecPaths(g)